Rename | dialog 폴더 생성 및 파일 위치 변경
