+++ app/src/main/java/com/example/plantogether/adapter/DateViewAdapter.kt
@@ -3,7 +3,7 @@ package com.example.plantogether.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.plantogether.data.EventData
+import com.example.plantogether.dialog.data.EventData
 import com.example.plantogether.databinding.RowAddscheduleBinding
 
 class DateViewAdapter(val items : ArrayList<EventData>): RecyclerView.Adapter<DateViewAdapter.ViewHolder>() {
@@ -32,7 +32,7 @@ class DateViewAdapter(val items : ArrayList<EventData>): RecyclerView.Adapter<Da
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.binding.eventNameText.setText(items[position].title)
+        holder.binding.planNameText.setText(items[position].title)
 
 
         holder.binding.eventList.setOnClickListener {
